+++ src/domain/challenge/challenge/routing/ChallengeRoute.tsx
@@ -6,7 +6,7 @@ import { useChallenge } from '../hooks/useChallenge';
 import { ApplicationTypeEnum } from '../../../community/application/constants/ApplicationType';
 import { PageProps } from '../../../shared/types/PageProps';
 import { Error404 } from '../../../../core/pages/Errors/Error404';
-import ApplyRoute from '../../../community/application/routing/apply.route';
+import ApplyRoute from '../../../community/application/routing/ApplyRoute';
 import { nameOfUrl } from '../../../../core/routing/urlParams';
 import { OpportunityProvider } from '../../opportunity/context/OpportunityProvider';
 import { CommunityContextProvider } from '../../../community/community/CommunityContext';
@@ -19,6 +19,7 @@ import CalloutRoute from '../../../collaboration/callout/routing/CalloutRoute';
 import ChallengeContextPage from '../pages/ChallengeContextPage';
 import ChallengeOpportunityPage from '../pages/ChallengeOpportunityPage';
 import CalloutsPage from '../../../collaboration/callout/CalloutsPage';
+import ChallengePageLayout from '../layout/ChallengePageLayout';
 
 interface ChallengeRootProps extends PageProps {}
 
@@ -69,7 +70,16 @@ const ChallengeRoute: FC<ChallengeRootProps> = ({ paths: _paths }) => {
           }
         />
       </Route>
-      <Route path={'apply/*'} element={<ApplyRoute paths={currentPaths} type={ApplicationTypeEnum.challenge} />} />
+      <Route
+        path={'apply/*'}
+        element={
+          <ApplyRoute
+            paths={currentPaths}
+            type={ApplicationTypeEnum.challenge}
+            entityPageLayout={ChallengePageLayout}
+          />
+        }
+      />
       <Route path={'feedback/*'} element={<CommunityFeedbackRoute paths={currentPaths} />} />
       <Route
         path={`${routes.Opportunities}/:${nameOfUrl.opportunityNameId}/*`}
