+++ src/domain/challenge/hub/routing/HubRoute.tsx
@@ -4,7 +4,7 @@ import { ChallengeProvider } from '../../challenge/context/ChallengeProvider';
 import { CommunityContextProvider } from '../../../community/community/CommunityContext';
 import { useHub } from '../HubContext/useHub';
 import { ApplicationTypeEnum } from '../../../community/application/constants/ApplicationType';
-import ApplyRoute from '../../../community/application/routing/apply.route';
+import ApplyRoute from '../../../community/application/routing/ApplyRoute';
 import { nameOfUrl } from '../../../../core/routing/urlParams';
 import ChallengeRoute from '../../challenge/routing/ChallengeRoute';
 import { PageProps } from '../../../shared/types/PageProps';
@@ -16,6 +16,7 @@ import CalloutRoute from '../../../collaboration/callout/routing/CalloutRoute';
 import HubContextPage from '../pages/HubContextPage';
 import HubDashboardPage from '../pages/HubDashboardPage';
 import CalloutsPage from '../../../collaboration/callout/CalloutsPage';
+import HubPageLayout from '../layout/HubPageLayout';
 
 export const HubRoute: FC<PageProps> = ({ paths: _paths }) => {
   const { displayName } = useHub();
@@ -50,7 +51,10 @@ export const HubRoute: FC<PageProps> = ({ paths: _paths }) => {
           element={<CalloutRoute parentPagePath={`${resolved.pathname}/${routes.Explore}`} entityTypeName="hub" />}
         />
       </Route>
-      <Route path={'apply'} element={<ApplyRoute paths={currentPaths} type={ApplicationTypeEnum.hub} />} />
+      <Route
+        path={'apply'}
+        element={<ApplyRoute paths={currentPaths} type={ApplicationTypeEnum.hub} entityPageLayout={HubPageLayout} />}
+      />
       <Route
         path={`challenges/:${nameOfUrl.challengeNameId}/*`}
         element={
